Added Animations
Animations will play based on velocity and facingDir
If either vel.x or vel.y are > 0, the respective Move animation will play.
func play_animation will only start playing an animation if it's not already playing. Allows it to loop correctly.
When idle, facingDir is used to determine which Idle frame to use.